refactor: align partition group ops
